refactor jaxrs-postgres-quarkus #392
