+++ application/jaxrs-postgres-quarkus/src/main/java/webapp/tier/healthcheck/LivenessHealthCheckPostgres.java
@@ -8,7 +8,7 @@ import org.eclipse.microprofile.health.Liveness;
 
 @Liveness
 @ApplicationScoped
-public class LivenessHealthCheck implements HealthCheck {
+public class LivenessHealthCheckPostgres implements HealthCheck {
 
 	@Override
 	public HealthCheckResponse call() {
